+++ R/crane-package.R
@@ -3,6 +3,7 @@
 #' @import ggplot2
 #' @import glue glue
 #' @importFrom broom tidy
+#' @importFrom broom.helpers tidy_parameters
 #' @importFrom cowplot plot_grid ggdraw draw_plot
 #' @importFrom dplyr across starts_with ends_with contains matches num_range
 #' @importFrom dplyr all_of any_of everything last_col where mutate